flaskServer/requirements.txt
pinecone-client==2.2.4
sentence_transformers==2.2.2
flask==2.3.3
langchain==0.0.291
openai==0.28.0